fall-detection-github/requirements.txt
opencv-python
mediapipe
scikit-learn
Pillow
numpy
requests
pandas
matplotlib
pygame
